Different implementation of DCGAN #3
Resolves #3
Description: Adapt a different implementation of the DCGAN, based on the tensorflow/tpu dcgan model.

Despite still getting the warning mentioned in #3, with the changes of this commit the model trains properly. I am still not sure then how the model worked in the first place (or more precisely, why it suddenly stopped working).

Disclaimer: The warning still appears, and I have no clue if this (and how) affects to anything during training.