Add encoder network #4
Closes #4

Adds the option to add an encoder netowrk to the model. To do so we use
the same discriminator function. It can be created either as an extension
attached at the end of the discriminator netowork (as it's done in InfoGAN)
or in a completely independent maneer. In the latter case a new scope
"Encoder" is passed to the Discriminator function; so the structure is
still the same, just with independent weights.

For now the training of the encoder affects the loss of the Generator in
any case. This option can be deactivated if desired.