Do param check for relevant parameters
The checking of the parameters is now done on the parameters that affect the model. Changes on non relevant parameters for the model/training, as the number of visualized images on predict or the TPU name, will be recorded but won't stop execution.

A flag to ignore the checking step and overwrite in any case has been added.