New dcm to tfrecord file
Add missing elements and output sole tfrecord file.

Description: The dcm array handling is not as trivial as we thought, so we just convert using the pydicom function, and we save the maximum values, as well as, a mask of the values outside of the scanner area.

Just producing a single .tfrecord file has only and exclusively I/O optimization motivations (which at it's time are economically motivated)
	( •͡˘ _•͡˘)ノð